Refactor instrumentation into InstrumentedHostnameVerifier
